refactor RecentProjects component; move to new file and update import path added new project
